+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - CLI: completely new `mc iam` command with support for user, roles, and user groups management
 - SDK: Tenant Management added paging to the Tenant Management Client
+- mc dev-proxy reads now --passkey from MDSP_PASSKEY variable like other commands [#301]
 - fixed paging issue for `mc iam --group` command [#297]
 - Bumped all dependencies